Add app.js
This is the first Draft of the functionality. As far as I know this should work
or at least provide a good start point to debug and develop a final solution.
It takes some values from the HTML document into variables and brings it to the
rest function that creates the JSON and the XMLHTTPRequest to send it.